SANTANDER/requirements.txt
pdfplumber==0.7.6
pandas==2.2.3
# Optional (OCR fallback). To enable OCR install these packages and system tools (Tesseract and Poppler):
pytesseract==0.3.10
pillow==10.0.0
pdf2image==1.16.3  # requires poppler on the system
